requirements.txt
boto3==1.9.188
botocore==1.12.188
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1
django-cors-headers==3.0.2
django-debug-toolbar==2.0
django-extensions==2.1.9
django-filter==2.1.0
django-filters==0.2.1
django-multiselectfield==0.1.8
django-reservation==0.2.9
django-rest-swagger==2.1.2
django-storages==1.7.1
django-tagging==0.4.6
django-taggit==1.1.0
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
djangorestframework==3.9.4
docutils==0.14
drf-yasg==1.16.1
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==19.0
Pillow==6.1.0
psycopg2-binary==2.8.3
pygraphviz==1.5
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
ruamel.yaml==0.15.99
s3transfer==0.2.1
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3